news-summarize/requirements.txt
# Core dependencies for news summarizer
requests>=2.31.0
beautifulsoup4>=4.12.0
newspaper3k>=0.2.8
feedparser>=6.0.10
openai>=1.0.0
python-dotenv>=1.0.0

# HTML parsing and cleaning
lxml>=4.9.0
lxml-html-clean>=0.4.2

# Additional dependencies that might be needed
nltk>=3.8.1
Pillow>=10.0.0
